requirements.txt
python-dotenv
streamlit
sentence-transformers
chromadb
requests
Biopython
fpdf2
pandas